Minor: changed return value to double
git-svn-id: https://visionegg.org/svn/trunk@1112 3a63a0ee-37fe-0310-a504-e92b6e0a3ba7

+++ visionegg/swig_src/win32_getrefresh.c
@@ -1,11 +1,11 @@
 #include <WINDOWS.H>
 
-int getrefresh() {
+double getrefresh() {
   DEVMODE DevMode;
 
   DevMode.dmDriverExtra = 0;
   if (EnumDisplaySettings(NULL,ENUM_CURRENT_SETTINGS,&DevMode)) {
-    return DevMode.dmDisplayFrequency;
+    return (double)DevMode.dmDisplayFrequency;
   } else {
     return 0;
   }
